+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
-## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [2.7.0]
+## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [2.8.0]
+This version of ONE adds support for loading .npz files.
+
+### Modified
+
+- one.alf.io.load_file_content loads npz files and returns only array if single compressed array with default name of 'arr_0'.
+- log warning when instantiating RegistrationClient with AlyxClient REST cache active
+- bugfix in load_collection when one or more files missing 
+
+## [2.7.0]
 This version of ONE adds support for Alyx 2.0.0 and pandas 3.0.0 with dataset QC filters. This version no longer supports 'data' search filter.
 
 ### Added